requirements.txt
python-nmap
colorama
fade
